requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
pyuca==1.2
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.8.0
